backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi
uvicorn
langchain-core
langchain-community
langchain-huggingface
langchain-groq
langchain-text-splitters
sentence-transformers
transformers
torch
pytesseract
pdf2image
pypdf
PyMuPDF
python-dotenv
python-multipart
pillow
beautifulsoup4
lxml
requests
supabase